pyproject.toml
[project]
name = "render"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "jax[cpu]>=0.4.13",
    "jaxtyping>=0.3.2",
    "typeguard>=4.4.4",
    "pyyaml>=6.0.2",
    "pillow>=11.3.0",
    "setuptools>=80.9.0",
    "optax>=0.2.5",
    "gradio>=5.43.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "plotly>=6.3.0",
    "tqdm>=4.67.1",
]

[tool.black]
skip-string-normalization = true
line-length = 100

[tool.ruff]
line-length = 100
